feat: Add jwt auth when user has succesfully logged in
if the User is not logged in it will not be able to acces the Chat Route, and once is logged in it will be redirected to products. If the user is logge din and try to navigate to /login or /signup it will also be redirected. Implemented /logout route to delete the stored jwt